修改了gimbal.c中视觉数据的发送逻辑，修改rm_referee.c中的crc校验为小写（兼容视觉）

+++ modules/referee/crc_ref.c
@@ -63,12 +63,12 @@ const uint16_t wCRC_Table[256] =
 
 // CRC8
 void append_crc8_check_sum( uint8_t *pchMessage, uint16_t dwLength);
-uint32_t verify_crc8_check_sum( uint8_t *pchMessage, uint16_t dwLength);
+uint8_t verify_crc8_check_sum( uint8_t *pchMessage, uint16_t dwLength);
 uint8_t get_crc8_check_sum( uint8_t *pchMessage, uint16_t dwLength, uint8_t ucCRC8 );
 
 // CRC16
 void append_crc16_check_sum(uint8_t * pchMessage,uint32_t dwLength);
-uint32_t verify_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength);
+uint8_t verify_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength);
 uint16_t get_crc16_check_sum(uint8_t *pchMessage,uint32_t dwLength,uint16_t wCRC);
 
 
@@ -94,7 +94,7 @@ uint8_t get_crc8_check_sum( uint8_t *pchMessage, uint16_t dwLength, uint8_t ucCR
 ** Input: Data to Verify,Stream length = Data + checksum
 ** Output: True or False (CRC Verify Result)
 */
-uint32_t verify_crc8_check_sum( uint8_t *pchMessage, uint16_t dwLength)
+uint8_t verify_crc8_check_sum( uint8_t *pchMessage, uint16_t dwLength)
 {
 		uint8_t ucExpected = 0;
 	
@@ -156,7 +156,7 @@ uint16_t get_crc16_check_sum(uint8_t *pchMessage,uint32_t dwLength,uint16_t wCRC
 ** Input: Data to Verify,Stream length = Data + checksum
 ** Output: True or False (CRC Verify Result)
 */
-uint32_t verify_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength)
+uint8_t verify_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength)
 {
 		uint16_t wExpected = 0;
