+++ modules/referee/crc_ref.h
@@ -9,12 +9,12 @@
 #define FALSE 0
 // CRC8
 void append_crc8_check_sum(uint8_t *pchMessage, uint16_t dwLength);
-uint32_t verify_crc8_check_sum(uint8_t *pchMessage, uint16_t dwLength);
+uint8_t verify_crc8_check_sum(uint8_t *pchMessage, uint16_t dwLength);
 uint8_t get_crc8_check_sum(uint8_t *pchMessage, uint16_t dwLength, uint8_t ucCRC8);
 
 // CRC16
 void append_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength);
-uint32_t verify_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength);
+uint8_t verify_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength);
 uint16_t get_crc16_check_sum(uint8_t *pchMessage, uint32_t dwLength, uint16_t wCRC);
 
 #endif
